+++ app/helpers/dataframe_handler.py
@@ -3,50 +3,11 @@ import logging
 
 # Modules
 import pandas as pd
-from json import dumps
 
 # Initialize log
 log = logging.getLogger(__name__)
 
-
-def parse_dataframe_columns_to_dictionary(dataframe: pd.DataFrame, dict_key: str, dict_value: str) -> dict:
-    '''
-    Read two dataframe columns and parse them to a dictonary.
-
-    Parameters
-    ----------
-    dataframe: pd.DataFrame
-        Dataframe to convert.
-    dict_key: str
-        Column name to be used as key for the dictonary
-    dict_value: str
-        Column name to be used as value for the dictonary
-
-    Returns
-    -------
-        dict
-            A dictionary with the key/value specified by the user.
-    '''
-    # Checking the dictonary key and value to ensure that the user input is found in the dataframe.
-    if dict_key not in dataframe:
-        raise ValueError(f'The column "{dict_key}" does not exist in the dataframe.')
-
-    if dict_value not in dataframe:
-        raise ValueError(f'The column "{dict_value}" does not exist in the dataframe.')
-
-    # Converting dataframe into a dictonary using user input to set key and value of the dictonary.
-    try:
-        dict_set = dataframe.set_index(dict_key).to_dict()[dict_value]
-        log.info(f'Dataframe was parsed to a dictonary with key: "{dict_key}" and value: "{dict_value}".')
-        log.debug(dumps(dict_set, indent=4, ensure_ascii=False))
-
-    except Exception as e:
-        log.exception(f'Creating dictonary from dataframe columns "{dict_key}" and "{dict_value}" failed with message: {e}')
-        raise
-
-    return dict_set
-
-
+# TODO: move to lib
 def verify_dataframe_columns(dataframe: pd.DataFrame, expected_columns: list, allow_extra_columns: bool = False) -> bool:
     """Verify if columns in dataframe contains expected colums.
