+++ app/helpers/join_data.py
@@ -1,0 +1,71 @@
+# Generic modules
+import logging
+
+# Modules
+import pandas as pd
+
+# Initialize log
+log = logging.getLogger(__name__)
+
+def create_dlr_dataframe(conductor_dataframe: pd.DataFrame,
+                         dd20_to_scada_name: dict,
+                         lineseg_to_mrid_dataframe: pd.DataFrame) -> pd.DataFrame:
+    """
+    TODO: doc me
+    """
+
+    # constant
+    MRIDMAP_DLR_ENABLED_COL_NM = 'DLR_ENABLED'
+    LINE_EMSNAME_COL_NM = 'LINE_EMSNAME'
+    # EXPECTED_NAME_COL_NM = 'ACLINE_EMSNAME_EXPECTED'
+    EXPECTED_NAME_COL_NM = 'acline_name_translated'
+
+    # append column with list mapped name based on expected name if is existing in list, else keep name.
+    # Remove expected name column
+    mapped_name_list = [dd20_to_scada_name[x] if x in dd20_to_scada_name else x
+                        for x in conductor_dataframe[EXPECTED_NAME_COL_NM]]
+    conductor_dataframe[LINE_EMSNAME_COL_NM] = mapped_name_list
+    conductor_dataframe = conductor_dataframe.drop(columns=[EXPECTED_NAME_COL_NM])
+
+    # TODO rename column instead or just keep translated name and make new column
+
+    # extract lists of unique line names from conductor and scada dataframe
+    # TODO: remove lines which are below 132 as DLR will not be enabled for them?
+    # TODO: list(set(self.topics_consumed_list) - set(self.topics_produced_list)) or differnence instead of comprehensions?
+    # TODO: names_not_in_gis = list(set(mrid_list).difference(translated_names))
+    # TODO: found_lines = list(set(mrid_list).intersection(translated_names))
+    lines_in_conductor_data = set(conductor_dataframe[LINE_EMSNAME_COL_NM].to_list())
+    lines_in_scada_data = set(lineseg_to_mrid_dataframe[LINE_EMSNAME_COL_NM].to_list())
+
+    # Create list of lines which have DLR enabled flag set True
+    lines_dlr_enabled = lineseg_to_mrid_dataframe.loc[lineseg_to_mrid_dataframe[MRIDMAP_DLR_ENABLED_COL_NM] == "YES", LINE_EMSNAME_COL_NM].to_list()
+
+    # report line names which are in DD20, but not ETS as info
+    lines_only_in_conductor_data = sorted([x for x in lines_in_conductor_data if x not in lines_in_scada_data])
+    if lines_only_in_conductor_data:
+        log.info(f"Line(s) with name(s): '{lines_only_in_conductor_data}' exists in conductor data but not in SCADA data.")
+
+    # report line names which are in ETS, but not DD20 as info
+    lines_only_in_scada_data = sorted([x for x in lines_in_scada_data if x not in lines_in_conductor_data])
+    if lines_only_in_scada_data:
+        log.info(f"Line(s) with name(s): '{lines_only_in_scada_data}' exists in SCADA data but not in conductor data.")
+
+    # report lines for which DLR is enabled, but date is not availiable in DD20 as errors
+    lines_dlr_enabled_data_missing = sorted([x for x in lines_dlr_enabled if x not in lines_in_conductor_data])
+    if lines_dlr_enabled_data_missing:
+        log.error(f"Line(s) with name(s): '{lines_dlr_enabled_data_missing}', are enabled for DLR but has no conductor data.")
+
+    # Join two dataframes where emsname commen key
+    # TODO: how to handle missing data? (dlr enabled but no conductor data)
+    dlr_dataframe = lineseg_to_mrid_dataframe.join(conductor_dataframe.set_index(LINE_EMSNAME_COL_NM),
+                                                   on=LINE_EMSNAME_COL_NM,
+                                                   how='inner')
+
+    # replace yes/no with true/false
+    dlr_dataframe.loc[dlr_dataframe[MRIDMAP_DLR_ENABLED_COL_NM] == "YES", MRIDMAP_DLR_ENABLED_COL_NM] = True
+    dlr_dataframe.loc[dlr_dataframe[MRIDMAP_DLR_ENABLED_COL_NM] == "NO", MRIDMAP_DLR_ENABLED_COL_NM] = False
+
+    # force lowercase
+    dlr_dataframe.columns = dlr_dataframe.columns.str.upper()
+
+    return dlr_dataframe
