--- app/voltagelevel_handler.py
@@ -1,45 +1,0 @@
-def convert_voltage_level_to_letter(voltage_level: int) -> str:
-    # TODO: make using regex instead?
-    """Converts voltage level to voltage letter representation.
-
-    Parameters
-    ----------
-    voltage_level : int
-        Voltage level in kV.
-
-    Returns
-    -------
-    str
-        Voltage letter.
-
-    Example
-    -------
-        >>> convert_voltage_level_to_letter(400)
-        C
-    """
-    if voltage_level >= 420:
-        voltage_letter = 'B'
-    elif 380 <= voltage_level < 420:
-        voltage_letter = 'C'
-    elif 220 <= voltage_level < 380:
-        voltage_letter = 'D'
-    elif 110 <= voltage_level < 220:
-        voltage_letter = 'E'
-    elif 60 <= voltage_level < 110:
-        voltage_letter = 'F'
-    elif 45 <= voltage_level < 60:
-        voltage_letter = 'G'
-    elif 30 <= voltage_level < 45:
-        voltage_letter = 'H'
-    elif 20 <= voltage_level < 30:
-        voltage_letter = 'J'
-    elif 10 <= voltage_level < 20:
-        voltage_letter = 'K'
-    elif 6 <= voltage_level < 10:
-        voltage_letter = 'L'
-    elif 1 <= voltage_level < 6:
-        voltage_letter = 'M'
-    elif voltage_level < 1:
-        voltage_letter = 'N'
-
-    return voltage_letter
